requirements.txt
python-dotenv~=1.1.0
aiogram~=3.20.0.post0
aiohttp~=3.11.18
feedparser~=6.0.11
bs4~=0.0.2
beautifulsoup4~=4.13.4
APScheduler~=3.11.0
pytz~=2025.2
httpx~=0.28.1
langdetect~=1.0.9
translate~=3.6.1